+++ gpio.h
@@ -2,6 +2,6 @@
 #define _GPIO_H_
 
 void update_switch_status (void);
-uint32_t get_switch_status (void);
+uint16_t get_switch_status (void);
 
 #endif
